(PC-17812) favorites/fixtures: rename favorisResponseSnap to favoritesResponse
