+++ src/features/favorites/fixtures/favoritesResponse.ts
