+++ src/features/favorites/pages/__tests__/useFavorites.test.tsx
@@ -8,7 +8,7 @@ import { useAuthContext } from 'features/auth/AuthContext'
 import {
   addFavoriteJsonResponseSnap,
   paginatedFavoritesResponseSnap,
-} from 'features/favorites/fixtures/favorisResponseSnap'
+} from 'features/favorites/fixtures/favoritesResponse'
 import { FavoritesWrapper } from 'features/favorites/pages/FavoritesWrapper'
 import { offerResponseSnap } from 'features/offer/api/snaps/offerResponseSnap'
 import { env } from 'libs/environment'
